+++ src/main.ts
@@ -28,11 +28,13 @@ process.env.UIX_BASE_PATH = process.env.UIX_BASE_PATH_OVERRIDE || resolve(__dirn
 async function bootstrap(): Promise<NestFastifyApplication> {
   const startupConfig = await getStartupConfig()
 
+  // (1) Create fastify adapter
   const fAdapter = new FastifyAdapter({
     https: startupConfig.httpsOptions,
     logger: startupConfig.debug || false,
   })
 
+  // (2) Register multipart with file size limit
   fAdapter.register(fastifyMultipart, {
     limits: {
       files: 1,
@@ -40,6 +42,7 @@ async function bootstrap(): Promise<NestFastifyApplication> {
     },
   })
 
+  // (3) Register helmet with custom CSP
   fAdapter.register(helmet, {
     hsts: false,
     frameguard: false,
@@ -71,6 +74,7 @@ async function bootstrap(): Promise<NestFastifyApplication> {
     },
   })
 
+  // (4) Create nest app with fastify adapter
   const app = await NestFactory.create<NestFastifyApplication>(
     AppModule,
     fAdapter,
@@ -92,7 +96,7 @@ async function bootstrap(): Promise<NestFastifyApplication> {
     res.send(await readFile(resolve(process.env.UIX_BASE_PATH, 'public/index.html')))
   })
 
-  // Serve static assets with a long cache timeout
+  // (7) Serve static assets with a long cache timeout
   app.useStaticAssets({
     root: resolve(process.env.UIX_BASE_PATH, 'public'),
     setHeaders(res) {
@@ -103,20 +107,20 @@ async function bootstrap(): Promise<NestFastifyApplication> {
   // Set prefix
   app.setGlobalPrefix('/api')
 
-  // Setup cors
+  // (9) Set up cors
   app.enableCors({
     origin: ['http://localhost:8080', 'http://localhost:4200'],
     methods: ['GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'OPTIONS'],
   })
 
-  // Validation pipes
+  // (10) Set up validation pipes for the api
   // https://github.com/typestack/class-validator
   app.useGlobalPipes(new ValidationPipe({
     whitelist: true,
     skipMissingProperties: true,
   }))
 
-  // Setup swagger api doc generator
+  // (11) Build and serve swagger api docs at /swagger
   const options = new DocumentBuilder()
     .setTitle('Homebridge UI API Reference')
     .setVersion(configService.package.version)
@@ -130,13 +134,13 @@ async function bootstrap(): Promise<NestFastifyApplication> {
       },
     })
     .build()
-
   const document = SwaggerModule.createDocument(app, options)
   SwaggerModule.setup('swagger', app, document)
 
-  // Serve spa on all 404
+  // (12) Use the spa filter to serve index.html for any non-api routes
   app.useGlobalFilters(new SpaFilter())
 
+  // (13) Start listening - woohoo!
   logger.warn(`Homebridge UI v${configService.package.version} is listening on ${startupConfig.host} port ${configService.ui.port}.`)
   await app.listen(configService.ui.port, startupConfig.host)