+++ ui/src/app/modules/settings/settings.component.ts
@@ -7,6 +7,7 @@ import { FormControl, FormsModule, ReactiveFormsModule, UntypedFormControl } fro
 import { Router, RouterLink } from '@angular/router'
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
 import { TranslatePipe, TranslateService } from '@ngx-translate/core'
+import { isStandalonePWA } from 'is-standalone-pwa'
 import { ToastrService } from 'ngx-toastr'
 import { firstValueFrom } from 'rxjs'
 import { debounceTime } from 'rxjs/operators'
@@ -84,6 +85,7 @@ export class SettingsComponent implements OnInit {
   public runningOnRaspberryPi = this.$settings.env.runningOnRaspberryPi
   public platform = this.$settings.env.platform
   public enableTerminalAccess = this.$settings.env.enableTerminalAccess
+  public isPwa = Boolean(isStandalonePWA())
 
   public hbNameIsInvalid = false
   public hbNameIsSaving = false
@@ -211,6 +213,20 @@ export class SettingsComponent implements OnInit {
 
     await this.initNetworkingOptions()
     await this.initStartupSettings()
+
+    // Some settings might need to be disabled for some users
+    // (1) Disable some settings that can modify the URL from being changed from a PWA
+    //     This is to stop users from getting stuck if they change the port for example
+    if (this.isPwa) {
+      this.uiPortFormControl.disable()
+      this.uiHostFormControl.disable()
+      this.uiProxyHostFormControl.disable()
+      this.uiSslTypeFormControl.disable()
+      this.uiSslKeyFormControl.disable()
+      this.uiSslCertFormControl.disable()
+      this.uiSslPfxFormControl.disable()
+      this.uiSslPassphraseFormControl.disable()
+    }
 
     this.hbNameFormControl.patchValue(this.$settings.env.homebridgeInstanceName)
     this.hbNameFormControl.valueChanges